changed Subject for Observer implementation from Tests class to Invoker class
